moved public to src and added index.d.ts for types
